+++ geom/type_envelope.go
@@ -287,8 +287,8 @@ func (e Envelope) TransformXY(fn func(XY) XY) (Envelope, error) {
 	return NewEnvelope([]XY{fn(min), fn(max)})
 }
 
-// Box converts this Envelope to an rtree.Box.
-func (e Envelope) Box() (rtree.Box, bool) {
+// AsBox converts this Envelope to an rtree.Box.
+func (e Envelope) AsBox() (rtree.Box, bool) {
 	return rtree.Box{
 		MinX: e.minX(),
 		MinY: e.minY,